Add files via upload
Having some issues with Git bash, hence manual upload.
This is a programme that reads a lists from a file written as a text.I use the csv module to employ its library . It will print the values in each row. Features to remove the square bracket and change the format have been commented out but can be enabled for demonstration purposes.